+++ lib/cuprum/rails/responders/serialization.rb
@@ -1,24 +1,20 @@
 # frozen_string_literal: true
 
 require 'cuprum/rails/responders'
+require 'cuprum/rails/serializers/base_serializer'
+require 'cuprum/rails/serializers/context'
 
 module Cuprum::Rails::Responders
   # Implements serializing a result value into response data.
   module Serialization
-    # @param root_serializer [Class] The root serializer for serializing the
-    #   result value.
     # @param serializers [Hash<Class, Object>] The serializers for converting
     #   result values into serialized data.
     # @param options [Hash] Additional parameters for the responder.
-    def initialize(root_serializer:, serializers:, **options)
+    def initialize(serializers:, **options)
       super(**options)
 
-      @root_serializer = root_serializer
-      @serializers     = serializers
+      @serializers = serializers
     end
-
-    # @return [Object] the root serializer for serializing the result value.
-    attr_reader :root_serializer
 
     # @return [Hash<Class, Object>] The serializers for converting result values
     #   into serialized data.
@@ -30,7 +26,13 @@ module Cuprum::Rails::Responders
     #
     # @return [Object] the serialized data.
     def serialize(object)
-      root_serializer.call(object, serializers: serializers)
+      context = Cuprum::Rails::Serializers::Context.new(
+        serializers: serializers
+      )
+
+      Cuprum::Rails::Serializers::BaseSerializer
+        .instance
+        .call(object, context: context)
     end
   end
 end