+++ lib/cuprum/rails/serializers/json.rb
@@ -17,7 +17,6 @@ module Cuprum::Rails::Serializers
       'cuprum/rails/serializers/json/hash_serializer'
     autoload :IdentitySerializer,
       'cuprum/rails/serializers/json/identity_serializer'
-    autoload :Serializer, 'cuprum/rails/serializers/json/serializer'
 
     # Default serializers for handling value objects and data structures.
     #